+++ crates/ruma-macros/src/identifiers/id_dst/parse.rs
@@ -8,6 +8,9 @@ use syn::{meta::ParseNestedMeta, parse_quote};
 use super::{IdDst, OwnedId, Types};
 use crate::util::RumaCommon;
 
+/// The default size of the inline array for the `SmallVec` inner representation.
+const SMALLVEC_INLINE_BYTES_DEFAULT: usize = 32;
+
 impl IdDst {
     /// Parse the given `IdDst` macro input.
     pub(super) fn parse(input: syn::ItemStruct) -> syn::Result<Self> {
@@ -21,7 +24,7 @@ impl IdDst {
             attr.parse_nested_meta(|meta| id_dst_attrs.try_merge(meta, attr))?;
         }
 
-        let IdDstAttrs { validate } = id_dst_attrs;
+        let IdDstAttrs { validate, smallvec_inline_bytes } = id_dst_attrs;
 
         if validate.is_none() && !input.generics.params.is_empty() {
             return Err(syn::Error::new(
@@ -52,6 +55,8 @@ impl IdDst {
         })?
         .into();
 
+        let smallvec_inline_bytes = smallvec_inline_bytes.unwrap_or(SMALLVEC_INLINE_BYTES_DEFAULT);
+
         let generics = input.generics;
         let (impl_generics, type_generics, _where_clause) = generics.split_for_impl();
         let impl_generics = quote! { #impl_generics };
@@ -61,9 +66,9 @@ impl IdDst {
         let owned_ident = format_ident!("Owned{ident}");
         let owned_id_type = parse_quote! { #owned_ident #type_generics };
 
-        let owned_id = OwnedId::new(owned_ident, owned_id_type);
-
+        let owned_id = OwnedId::new(owned_ident, owned_id_type, smallvec_inline_bytes);
         let ruma_common = RumaCommon::new();
+        let types = Types::new(&ruma_common, &owned_id);
 
         Ok(Self {
             ident,
@@ -73,7 +78,7 @@ impl IdDst {
             validate,
             str_field_index,
             owned_id,
-            types: Types::new(&ruma_common),
+            types,
             ruma_common,
         })
     }
@@ -84,6 +89,9 @@ impl IdDst {
 struct IdDstAttrs {
     /// The path to the function to use to validate the identifier.
     validate: Option<syn::Path>,
+
+    /// The size of the inline array for the `SmallVec` inner representation.
+    smallvec_inline_bytes: Option<usize>,
 }
 
 impl IdDstAttrs {
@@ -102,6 +110,25 @@ impl IdDstAttrs {
         Ok(())
     }
 
+    /// Set the size of the inline array for the `SmallVec` inner representation.
+    ///
+    /// Returns an error if it is already set or if the value doesn't fit into a `usize`.
+    fn set_smallvec_inline_bytes(
+        &mut self,
+        inline_bytes: syn::LitInt,
+        attr: &syn::Attribute,
+    ) -> syn::Result<()> {
+        if self.smallvec_inline_bytes.is_some() {
+            return Err(syn::Error::new_spanned(
+                attr,
+                "cannot have multiple values for `smallvec_inline_bytes` attribute",
+            ));
+        }
+
+        self.smallvec_inline_bytes = Some(inline_bytes.base10_parse()?);
+        Ok(())
+    }
+
     /// Try to parse the given meta item and merge it into this `IdDstAttrs`.
     ///
     /// Returns an error if an unknown `ruma_id` attribute is encountered, or if an attribute
@@ -109,6 +136,10 @@ impl IdDstAttrs {
     fn try_merge(&mut self, meta: ParseNestedMeta<'_>, attr: &syn::Attribute) -> syn::Result<()> {
         if meta.path.is_ident("validate") {
             return self.set_validate(meta.value()?.parse()?, attr);
+        }
+
+        if meta.path.is_ident("smallvec_inline_bytes") {
+            return self.set_smallvec_inline_bytes(meta.value()?.parse()?, attr);
         }
 
         Err(meta.error("unsupported `ruma_id` attribute"))